refactor: update entity folder layout for clarity
