+++ src/entities/base_entity.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "../../combat/health.h"
-#include "../../combat/hitbox.h"
-#include "../../core/direction.h"
+#include "../combat/health.h"
+#include "../combat/hitbox.h"
+#include "direction.h"
 #include <SFML/Graphics.hpp>
 #include <vector>
