+++ src/entities/enemies/race_condition_slime/states/chase_state.h
@@ -1,19 +1,16 @@
 #pragma once
-#include "../../base/enemy_state.h"
+#include "../../enemy_state.h"
 
 namespace rc_slime {
 
-// Slime is recovering after an attack. Stationary; transitions to Chase
-// if player still in range or back to Idle.
-class RecoverState : public EnemyState {
+// Slime actively pursues the player. Handles horizontal movement and jumping
+// Transitions to WindUp (if in attack range) or Idle (if player escapes).
+class ChaseState : public EnemyState {
   public:
 	[[nodiscard]] EnemyState *update(float deltaTime, BaseEnemy &enemy, const World &world,
 	                                 sf::Vector2f playerPos) override;
 	void updateAnimation(float deltaTime, BaseEnemy &enemy) override;
 	void onEnter(BaseEnemy &enemy) override;
-
-  private:
-	float stateTimer = 0.f;
 };
 
 } // namespace rc_slime
