+++ src/entities/enemies/race_condition_slime/states/idle_state.cpp
