+++ src/entities/player/abilities/hat_ability.cpp
@@ -1,5 +1,5 @@
 #include "hat_ability.h"
-#include "../../core/audio_manager.h"
+#include "../../../core/audio_manager.h"
 #include <cassert>
 
 HatAbility::HatAbility()
