+++ src/entities/player/abilities/hat_projectile.cpp
